+++ lace/examples/bench.rs
@@ -1,9 +1,11 @@
 use std::env;
 
+#[cfg(feature = "bencher")]
 use lace::bencher::Bencher;
 use lace::prelude::*;
 use lace_utils::{mean, std};
 
+#[cfg(feature = "bencher")]
 fn main() {
     let args: Vec<String> = env::args().skip(1).collect();
 
@@ -51,4 +53,9 @@ fn main() {
     let std_time = std(&times_sec);
 
     println!("Time (M, S) = ({}s, {}s)", mean_time, std_time);
+}
+
+#[cfg(not(feature = "bencher"))]
+fn main() {
+    eprintln!("Please enable the 'bencher' feature to use this example.")
 }